VariablesTableByStudy displays data passed from parent
